Xファイル関連クラスのUTF-8化
